requirements-docker.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
captum==0.8.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coverage==7.9.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.97.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.33.1
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
fastapi
uvicorn[standard]
python-multipart